pyproject.toml
[project]
name = "bbi"
version = "0.1.0"
description = 'Bounding-box inference agent with model-based planning and Q-learning.'
authors = [{name = "Lucas Cruz", email = "lucas@lucas-cruz.com"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT License"}
dependencies = [
    "numpy>=1.26.4",
    "gymnasium[classic-control]>=0.29.1",
    "goright[extra]@git+https://github.com/cruz-lucas/goright.git",
    "gin-config>=0.5.0",
    "structlog>=25.1.0",
    "mlflow>=2.22.0",
]

[project.urls]
Homepage = "https://lucas-cruz.com"
Documentation = "https://lucas-cruz.com"
Repository = "https://github.com/cruz-lucas/Bounding-Box-Inference.git"

[tool.uv.pip]
extra-index-url = []

[project.optional-dependencies]
dev = [
    "pre-commit>=4.0.1",
    "isort>=5.13.2",
    "mypy>=1.8.0",
    "ruff>=0.3.4",
    "black>=24.3.0"
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
strict = true
exclude = "mlruns"

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B"]
ignore = ["E501"]
